requirements.txt
Django
djangorestframework
mysqlclient
djangorestframework-simplejwt
python-decouple
django-cors-headers